refactor: extract ray pool executor
